+++ lib/src/app/usecase/default_crud_usecase_async.dart
@@ -22,8 +22,11 @@ abstract class DefaultCRUDUseCaseAsync<Domain extends BasicDomainObject>
   Future<Domain> findBy(int keyId) async => await repo.findBy(keyId);
 
   @override
-  Future<Domain> destroy(Domain objectToDestroy) async =>
+  Future<void> destroy(Domain objectToDestroy) async =>
       await repo.destroy(objectToDestroy);
+
+  @override
+  Future<void> destroyById(int id) async => await repo.destroyById(id);
 
   @override
   Future<void> init() async {}