+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * 1.2.4-RELEASE.20220000:
     * **IMPROVEMENT** :raised_hands: : Removed `DefaultGeneralConverter`, use instead `GeneralConverter`.
     * **NEW** :+: : Created architecture for Async operations.
+    * **NEW** :+: : Created JSONConverter for entities.
 
 * 1.2.3-RELEASE.20220206:
     * **BUG FIX** :bug: : Fix version number in Changelog.md